add a back button to parent page
This patch extends the template functions with back
button linking to the current pages' parent if
available. Both tpl_button() and tpl_actionlink()
are supported.

For this to work the first page in the namespace must
have the same name as the namespace itself. The 'back'
button of every page in this namespace links to
namespace:namespace. The 'back' button of the page
namespace:namespace links to the first page of the
containing namespace and so forth until the start page
has been reached.

Because of the precondition decribed above, the default
template hasn't got the new 'back' button. It is reserved
for custom made templates and installations which take
care of the precondition.

darcs-hash:20050519174025-7ef76-1e9c78c941f53871905fc1ba08b28a826553daa3.gz