keep search input on search page
This patch modifies the search input form so that the search input
is kept as long as the user stays on the search page. This is pure
comfort because he is able to optimise his query step by step without
the need of typing in the query multiple times.

darcs-hash:20050601163504-7ef76-aad10750707ba97810bc5c5185da3e8ad9ae7391.gz