extended search function
The DokuWiki search function uses 'OR' to combine
multiple search words. This behaviour is unusual and
not very helpful to narrow the search results.
This patch changed the behaviour to 'AND'. Multiple
search words will reduce the count of search results.

It uses assertions now. This has the big advantage
that only one regular expression have to be processed
for each file and the behaviour can be changed easily.

The functionallity has been extended: Words with a
preceding minus sign (-) will be excluded from and
words with a preceding plus sign (+) will be included
in the search results. Is a preceding sign is missing
(+) is assumed.

darcs-hash:20050603182746-7ef76-03b18a3707c91e8c5cbb99253b603d090b679c1a.gz