session_write_close added #364
I just learned that PHP does lock it's session objects. This is realy
bad if you have multiple images in a page as each one will call fetch.php
which locks the session, so everything can only be loaded sequentially.
The fix for this is to close the session after using it which is after doing
the auth and the breadcrumbs. I added the needed calls everywhere.

darcs-hash:20050608213514-9977f-f2dfc467ebbd06406e10e33a386fa1db6cb32dba.gz