mysql auth: added support for old passchecking method #359
This patch changes the mysql auth mechanism to support the old
method of password checking (leaving it to the DB) as well as
the new one. Which one is used is decided on which option is
defined:

$conf['auth']['mysql']['passcheck'] now behaves as in older
releases, You can use %u, %g and %p where %p contains the
cleartext password entered by the user. Access is granted
if the SQL statement returns one result row.

if $conf['auth']['mysql']['getpass'] is defined it is used
to fetch the crypted password from the database which is then
checked with auth_verifyPassword() - This is the preferred
method.

Users of the devel need to change their config by renaming
passcheck to getpass

darcs-hash:20050612111044-9977f-545feafc098082a067fdbbbc7d7d3a61c5903590.gz