pagetemplate support #104
This adds very simple pagetemplate support. When a file named _template.txt
exists in a namespace this file's contents are loaded into the editor when
a new page is added in the namespace.

Please note that this file is not editable from within the Wiki as it starts
with an underscore. The wiki admin has to create it her self on the
filesystem.

darcs-hash:20050718173842-7ad00-39b1b76bef4c6243fc00d94673793d68b356aecb.gz