caching for feeds
This patch adds caching for XML/ATOM feeds. This is most useful for the
default feed (linked from the footer and HTML meta header) as this is
usually the most used one.

The cache is invalidated if a change was made (by checking the the
filemtime of the changelog). A 5 minute settling time is used to avoid
invalidating the cache too often when the wiki changes rapidly.

darcs-hash:20050806155857-7ad00-5b9cb05b42d67c3f23f0358676735038b9f59664.gz