hidepages configoption
This new option accepts a RegExp to filter certain pages from all automatic
listings (RSS, recent changes, search results, index). This is useful to
exclude certain pages like the ones used in the sitebar templates. The
regexp is matched against the full page ID with a leading colon. If it
matches the page is assumed to be a hidden one.

IMPORTANT: this is not related to ACL. A hidden page is still visible to all
users (if not restricted by ACL) when linked or called directly.

darcs-hash:20051103101726-6e07b-8d45912a1b4f6cfc9e3fce147c15f84a58ea7ca2.gz