Added Google sitemap support #371
This patch addes the automatic creation of Google sitemaps. The map
is created in the DokuWiki root dir and named sitemap.xml.gz if gzip
compression is available - if not the gz extion is skipped.
How often the map is recreated is defined through the $conf['sitemap']
option. It accepts a day value.

darcs-hash:20051126234709-7ad00-6ff4b0e79670cdfa39e615ec9dc40146ffcc9dd4.gz