usermanager start value missing
The start value of the user list was corrupted, if in
search mode but _REQUEST['search'] was empty. An empty
start value passed to the authentication backends might
cause an error (in MySQL backend for sure). This patch
fixes this.

darcs-hash:20060126195356-7ef76-d899fb08bba42de487f695536f4ae0cdeb3a0eaa.gz