create unique IDs for sections
This patch finally completes the support for unique section IDs.

To achive this the mechanism how the TOC is build was changed. The
TOC now is build in the renderer only. Currently the TOC will be
rendered in the end_document function and is then prepended to the
doc. This should ensure compatibility with the rest of the code.

Adding support for separating the TOC from the page should now be a
simpler task in the future.

TODO:
  - Update base class
  - remove commented old TOC code
  - make sure no other parts of the code use any of the old TOC
    code

darcs-hash:20060218230744-7ad00-40c5463d93f8ae1c543fb4fed747b2047b4c1302.gz