event ACTION_TEMPLATE
This event is advisory only.
Event data is the $ACT variable
The default action is to display an "unknown action" message

The event is triggered by tpl_content if it does not recognise the $ACT value. It is
the complement to event 'ACTION_DISPATCH'.  Handlers of this event can output XHTML
which will be displayed in the main dokuwiki window - ie. where the wiki page would
normally be displayed.

darcs-hash:20060414201705-9b6ab-f8ede6a2fb601d734c115ee09bb57195ad7d1e2b.gz