config plugin ui update 20060520
This patch hides settings that are missing config metadata and optionally
provides a list of warnings about settings that are not properly configured.
  - Warnings about settings are listed if $conf['allowdebug'] is true.
  - Warnings are listed by the $conf string as it appears in local.php.
  - Warnings show the $meta string as it would appear in the correct
    settings metadata file.
  - There are 3 kinds of warnings.
    - undefined
      There is no $meta information defined for this setting.
    - no class
      The setting class specified in $meta can not be found.
      This setting does have a $meta entry.
    - no default
      The setting is missing a default value.
      The setting does have a $meta entry with a valid setting class.
  - Note: Settings with metadata but other warnings are allowed to appear
          in the normal config settings list.

Also...
  - Templates can now define their own settings classes.
  - Removed an XHTML validation error from the first patch.
  - More language strings to go with the new warnings.

The warnings under the 'Undefined Settings' heading are intended to
provide developers with a list of any settings that they have forgotten to
finish preparing for the config plugin. This list should be blank for stable
releases.

darcs-hash:20060520103718-05dcb-6d4e6bce78498cbf9d087e27d52e4aa30917b0a5.gz