decouple section edit from header in renderer
This patch achieves a partial decoupling of the section edit buttons
from the header syntax. A new 'section_edit' instruction is added to the
instruction list. The default set of these instructions is generated from
the header handler and the _finalize function. A generic 'section_edit'
renderer is now part of the renderer.

To rewrite the section semantics, add, or remove section edit buttons the
header handler and renderer no longer need to be modified. Changes can be
dynamically made to the section edit buttons through the 'PARSER_HANDLER_DONE'
action event. An action plugin that registers this event can then modify
the calls list of the handler object before it is rendered.

In short, this patch makes it possible to change the behavior of the section
edit buttons without having to change unrelated code. This patch should be
entirely transparent to DokuWiki users and cleanly written plugins.

darcs-hash:20060522060004-05dcb-8e6b5215e1b610877f126ca0b9e1e5bc21cc6469.gz