MySQL Bug - PHP warning if query result is empty
If a database query has no result, the returned array is not
valid. This caused a PHP warning in the user manager if the
filter is set in a way that the result list is empty.

darcs-hash:20060522162327-7ef76-d8fa89a047a9a159b600f11ccb189e845b2a5b8f.gz