be clever about extensions on upload
This patch makes it possible to enter filenames without an
extension in the upload as field when the uploaded file it
self has a known extension.

It also prints a warning when an extension was changed by
the user.

darcs-hash:20060527100310-7ad00-cf2abb4b9730e54b96c6ba24b2a999011cfb834a.gz