namespace linking first part
This patch adds namespace linking - formerly known as globalstart patch.

It differs somewhat from the original patch. It is not implemented in getID()
but in resolve_pageid().

It is now possible to link to a "default" file of a namespace ending the linkid
with a colon: [[foo:bar:]]

To which page the link links is dependent on the xistance of certain named
files. For the above mentioned link [[foo:bar:]] the following pages are checked:

foo:bar:start
foo:bar:bar
foo:bar

The pages are checked in the order above whatever page is found first will be
linked to. If no page is found foo:bar:start will be chosen. BTW: 'start' is
the value configured in $conf['start']

Note: autoplural linking is not done for those links

This is just the first patch. Several other locations of the code need to be
adjusted to reflect this change and some testing needs to be done (first test
cases are supplied within the patch bundle)

Things that maybe need adjustment:

  - tpl_youarehere (hierarchical breadcrumbs)
  - tpl_button back
  - maybe others (search?)

Patches would be welcome. The best aproach to fix things that don't work is
probably making calls to resolve_pageid() instead of simple cleanID() calls.

darcs-hash:20060611184453-7ad00-ba70b0fcf2cb64d4d4f0ce6bd6e437595cd947d3.gz