$conf['fetchsize'] added
This patch adds an option to configure the maximum size for files the fetch.php
will ever download. Setting this to 0 completely turns of the caching of external
media files.

Disadvantages of setting a low or zero fetchsize:

  * fetch.php needs to download images to be able to resize them. When the used
    fetchsize prevents the downloading the images can only be resized by the
    browser which means the browser will need to download the fullsized image first.

  * If the linked external media files vanishes it will no longer display in the
    wiki because it is not cached.

Advantages of setting a low or zero fetchsize:

  * fetch.php may be used for a possible denial of service attack by requesting
    many big external files.

  * The created cache files may take a lot of space on the server

I recommend to leave the setting at 2MB for internal and private wikis and lower
the setting to about 200 to 500 Kb for bigger public Wikis.

Note: the caching of files uploaded through the media manager is not affected by
this setting.

darcs-hash:20060615184847-7ad00-04fc39928f7d72e56f5c5e271013ef265436e6c9.gz