disabled actions addition to config plugin
This patch adds a new generic settings class, setting_multicheckbox.  The class
displays each of a list of possible choices as a checkbox and an additional text
input field for "other" values.  The final result is saved in the config file as
a comma separated list.

It also supports an additional, optional, parameter, "_combine" which can be used
to group complimentary values, allowing them to be set or unset together with one
checkbox.

darcs-hash:20060705012142-9b6ab-b92daddd1393bfc5e9b650a57348237726957c71.gz