IO action events
Adds page and namespace events:
  IO_WIKIPAGE_READ
  IO_WIKIPAGE_WRITE
  IO_NAMESPACE_CREATED
  IO_NAMESPACE_DELETED

The namespace events are purely advisory,
while the wikipage events allow page content
to be modified between DokuWiki and the disk.

These events are primarily intended to simplify
keeping other tools in sync with the semantic
structure of a DokuWiki site. As an added benefit,
the events allow plugins to conduct automated
processing of raw wiki page content.

The namespace events cover the separate namespace
trees for both pages and media. The "name" of the
tree that the event belongs to is included in the
event data.

darcs-hash:20060705105652-05dcb-f44024e852a2adf1a14b8a7d69c46db067e72307.gz