fix recent changes cache ordering
This patch fixes a bug in indexer.php which resulted in the order
of the recent changes cache being reversed each time it was trimmed.

It also adds sorting to both getRecents() and runTrimRecentChanges()
as a defensive measure against the order of the file being corrupted.

darcs-hash:20060923235109-9b6ab-0f4062c1b02449cce9382426174cd22d71387e5a.gz