rss_summary_config
Add a configuration option to control whether or not the edit/revision summary is
displayed with the title of RSS feeds. The default is to show the summary, the same as
DokuWiki did prior to this addition.

darcs-hash:20061101023313-9039d-d067e9bcd7dc0c7cabd57674d1072c5b882ad4e7.gz