added getFormat() function to renderer
Each renderer has to supply a getFormat() function returning the format it
produces. Usually this is the same as the classname of the renderer (or Plugin)
says but it does not need to be necessarily.

Fixes a problem with the wrong format being reported to plugins when a
render plugin was used.

darcs-hash:20070303214102-7ad00-3f87d8a10e94f4c6c30bacc8f43486b94ecdb146.gz