added JsStrip unit test files
This patch adds unit test files from the new release of JsStrip [1]. Some of the tests
currently fail pointing to bugs which are fixed in the new JsStrip release. The fixes
need to be ported to DokuWiki's js_compress function.

This patch also adds a way to output additional info when a test fails. It misuses
SimpleTest's signalling which is probably a bad idea but works for now.

http://code.google.com/p/jsstrip/

darcs-hash:20070305223025-7ad00-583a70a1ddc374eb61193f9315af99baeedea0be.gz