remove htmlok and phpok tests from Doku_Handler
there is no necessity for the "htmlok"
and "phpok" testing in the Doku_Handler, taken the fact that the page
renderer does it too. and it gives the possibility to the renderer to
choose, what to do with such blocks.

darcs-hash:20080117222648-01292-cfce862d4e6fe9ce61feda7d6f11f014a68f5ffa.gz