UNDO: remove htmlok and phpok tests from Doku_Handler
there is no necessity for the "htmlok"
and "phpok" testing in the Doku_Handler, taken the fact that the page
renderer does it too. and it gives the possibility to the renderer to
choose, what to do with such blocks.

darcs-hash:20080117222648-01292-c4b5a890ada5e5e962ba7e3f62c4c2fc5d72138f.gz