Change handler to use a defined character for line-break replacement
This patch is the first in a series aimed at making it easier for DW to allow
plugins to modify the standard handling of line-breaks.

Currently the handler "eats" line-break characters replacing them with a space.
This patch changes the replacement character to a define, DOKU_PARSER_EOL and
updates all the test cases to use the new define.

DOKU_PARSER_EOL is left as a space.

darcs-hash:20080309201116-f07c6-21edb4f6955da8fa5ba0f9346d859a7c5c69a239.gz