allow dynamic passing of template to use for css.php
This patch makes it possible to pass the template name to use to the
lib/exe/css.php dispatcher. When passed the $conf['template'] option is
ignored by the disaptcher and the given template is used instead.

This makes it possible to switch templates dynamically without loosing the
CSS dispatcher functionality. This might be useful for things like the
multitemplate template or for loading a template based on the user agent.

darcs-hash:20080310201630-7ad00-2062fa939b1f868540031ea42a42e948dd82bbb4.gz