change line-break replacement character to '\n'
This patch is the third in a series aimed at making it easier for DW to allow
plugins to modify the standard handling of line-breaks.

It effectively leaves the new-line character '\n' in the 'cdata' instruction
allowing the renderer to process the character as it chooses. No changes have
been made to the renderers.

This change is neutral to xhtml renderers as xhtml treats spaces and new-lines
identically (as white-space).

Writers of renderer plugins are advised to check their plugin against the new
patch.

darcs-hash:20080312000924-f07c6-a6dab2d16c7bc42e9dc8eb137648c6f984b4f2e6.gz