Add support for plugin renderers to replace standard renderers
This patch is the fourth in a series aimed at making it easier for DW to allow
plugins to modify the standard handling of line-breaks.

It adds:
- new config setting 'renderer_xhtml', default value 'xhtml'
- new renderer method 'reset()' which is used by reusable renderers when used
  to render second and subsequent data. (*)

An extra step has been added to the renderer selection process.
- check $conf["renderer_$mode] for renderer name.  If it doesn't exist use $mode as
  the renderer name.

(*) It maybe that the 'document_start()' method can be used for this.  However the
current xhtml does not correctly reset itself for reuse.

darcs-hash:20080312005647-f07c6-ff2cb960c05927f5c6f3e916a364fcad470c2ce3.gz