jump to edited section after editing FS#643
This patch adds a mecanism that tries to scroll down to the section that was
just edited when using section editing. The approach used is not failsafe, but
should work in most cases. It will not work correctly when:

  * multiple sections on the page have the same headline
  * the edited section no longer starts with a headline

darcs-hash:20080506205011-7ad00-9d213d4436c4f01840d3f62277c3fee9e316d447.gz