display the (shortened) namespace for page quicksearch
When displaying pagename matches in the Ajax quick search or normal search,
the namespace of the page is shown behind the pagename. This makes it easier
to distinguish the pages when the same namespace is used in different namespaces.

To avoid breaking the layout in deep nested namespace hierarchies, the namespace
is shortened in the middle when needed.

This patch also disables the effect of the useheadings option in the Ajax quick
search. After all the results should show what was found and since the search
works on pagenames not headings it should show pagenames as result.

darcs-hash:20080814194826-7ad00-9add9c1bbbb4f4ede3c6884d37427644b2cddc56.gz