Fixed doubleclick-behaviour of links, fixes FS#1389
There is a new open-attribute of listitems that represents the state that should be achieved or the current state. This makes it possible that clicks that occur when the opening of the listitem has already been triggered close it again and the listitem won't be reopened when the ajax-call is finished.

darcs-hash:20081012141018-074e0-f9be8bbe89083df761d71f665b409bc95e450bf3.gz