Make content-disposition configurable in mime.conf FS#1541
With this patch it is possible to define if a file should be served as download
or be displayed inside the browser (if supported) by configuring it in
conf/mime.conf

Mimetypes that should be served with a "Content-Disposition: attachment" header
need to be prefixed with a exclamation mark. All others will be served inline.

This will also fix a Problem with Flash 10.

darcs-hash:20081202210322-7ad00-6e7ef30aff9322cd135311be77809187da121f3b.gz