new headline mechanism in the toolbar
Ignore-this: f87882b727bcf6ea067945eea0f158b1

The toolbar now features three new buttons to insert a headline. Instead of
asking the user for the headline level it asks if she wants a higher, lower
or equally leveled section (relative to the current one).

This means the user no longer has to count in which section she currently is
and encourages users to use the proper order of headline levels.

To make room in the toolbar, the old buttons have been moved in a picker.

darcs-hash:20090528155327-7ad00-e384cd7e45cb192657e29c875f46e3676d47f367.gz