Fixed sorting in the acl manager treeview
Previously the content of expanded namespaces was displayed below
all other pages, now it is placed below it's parent namespace.
The new comparison function is quite complex. Please have a look at it
if it can done easier or if there are cases I haven't considered where
the function produces incorrect results.