Changes to the ft_pageLookup and related event FS#1978
This patch changes the ft_pageLookup function to always return the
title of pages with the result. This makes it easier to work with the
array, as it no longer changes between numeric and key indexes depending
on useheading.
This also means that action plugins subscribed to
SEARCH_QUERY_PAGELOOKUP need to be adjusted. The event contains a new
data field called 'has_titles' which plugins can use to check for
backwards compatibility.