Indexer speed improvement: joined array vs. single lines
From my experience with a benchmark of the indexer it is faster to first
join the array of all index entries and then write them back together
instead of writing every single entry. This might increase memory usage,
but I couldn't see a significant increase and this function is also only
used for the small index files, not for the large pagewords index.