Indexer improvement: regex instead of arrays for lines
When updating a single line that line was split into an array and in a
loop over that array one entry was removed and afterwards a new one
added. Tests have shown that using a regex for doing that is much faster
which can be easily explained as that regex is very simple to match
while a loop over an array isn't that fast. As that update function is
called for every word in a page the impact of this change is
significant.