Make baseonly work in allowedModes
This makes it possible to have modes that do accept headers
(baseonly) in the parser.

Related test cases are still running through but I'm not 100% sure
I did not break something here. So it should be tested a bit more.

This patch will allow plugins to wrap multiple sections, however it
also makes it possible to easily break XHTML validity, because
headers also open and close sections, so plugin authors need to be
aware!

In case you wonder: this patch is not about allowing formatting
inside headers.