Use md5 of the id in the subscription lockdir. FS#2112
Filenames can't contain ":" on windows and the lock directory contained
the unescaped page id. The lock function tries in an endless loop to
create the lock directory when it fails and the directory doesn't exist.
Just escaping the directory name won't work as then the filename length
limit will be quickly hit when using deep namespace structures with utf8
names. Thus using the md5sum seems to be the best solution. Perhaps the
lock function could also be changed to create a file with that name that
contains the id so the id can be retrieved for debugging purposes.