Use title index for more than 11 p_first_heading calls
This change makes p_get_first_heading load the title index when more
than 11 requests that caused a call to p_get_metadata have already been
done. This means that small pages and the breadcrums won't trigger the
loading of the title index but for larger pages or the sitemap the title
index will be used. This is necessary because every call to
p_get_metadata can trigger the parsing and rendering of a whole page and
there can be many calls when useheading is activated and e.g. the
index/sitemap page is displayed.

Additionally this adds a small title cache that caches titles requested
from p_get_metadata.

Further tests should be done how this affects memory usage and how often
the index loading is triggered in order to see if that parameter should
be adjusted.