Indexer v3 Rewrite: fix addMetaKeys and locking
This fixes addMetaKeys so it actually removes values. This also changes
the functionality of the function: It now updates the key for the page
with the current value instead of adding new values as this will be the
default use case. A new parameter could be added to restore the "old"
behavior when needed.
addMetaKeys now only saves the index when the content has really been
changed.
Furthermore no empty number is added anymore to the reverse index when
it has been empty previously.
addMetaKeys now releases the lock again and really fails when the lock
can't be gained.