allow placeholders in search intro text
This patch allows to use the placeholders in the search intro message that
will be replaced with the search term.

@SEARCH@ will be replaced with the search query
@QUERY@  will be replaced with the URL encoded search query for use in URL
         parameters

Please note that the replacement is don't on the XHTML *after* parsing and
rendering the intro wiki text. This means you can not use the query where
an ID would be expected.

Examples:

  This will work:
    [[http://www.google.com/search?q=@QUERY@|Google for @SEARCH@]].

  This will not work and will link to the page "search" instead:
    [[@SEARCH@|Your page]].

  You could use this instead:
    [[this>doku.php?id=@QUERY@|Your page]].